% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ralphloop-package.R
\docType{package}
\name{ralphloop-package}
\alias{ralphloop}
\alias{ralphloop-package}
\title{ralphloop: Persistent Iterative Loop for LLM-Driven Development}
\description{
Provides a persistent, truth-based iterative development loop for
large language models using the ellmer package.
}
\author{
\strong{Maintainer}: Lazarus Alvarez \email{lazarusalvarez3@gmail.com}

}
